add signdata
e4/a3d2faaa3e925e33b60f660508f9e6c83a645a63a62037995dfeda417dca71
is for ubuntu jammy secureboot livecd.

Note this doesn't quite follow the recommended best practice, in
that this is - for demonstration purposes only - re-using the
same pcr7 for provisioning and install.  IOW there is no limited
or tpmpass, only 'production'.  That's ok, this is only for
demonstration.

We use the 'snakeoil' keyclass since we want to use the
snakeoil keys for signing policies.

Signed-off-by: Serge Hallyn <serge@hallyn.com>