Provision disks
When doing 'trust provision', (optionally) create two disk partitions:

1. PBF, the plaintext partition, contains the pcr7 signature data
2. SBF, a luks-encrypted partition, contains an empty filesystem
   for future use.

The disk to use is specified as the --disk= argument to 'provision'.
If not specified, then no partitions will be created.  If it is
"any", then 'trust' will choose a disk for you.  If it is 'sdb' or
'/dev/sdb', then it will use sdb.

The chosen disk must be empty.  If the --wipe argument is given,
then the chosen disk will be wiped.  (be careful using this with
--disk=any)

Furthermore, since we now have PBF with signdata after provisioning,
try to bind mount PBF/pcr7data onto /pcr7data during early boot.

Note - it would probably be nicer to not require pbf and sbf be
partitions 1 and 2, to look for free space instead.

Signed-off-by: Serge Hallyn <serge@hallyn.com>