Add gen-sudi sub-command to trust.
The 'gen-sudi' subcommand takes a serial, product uuid, and uuid and
generates a sudo certificate and private key.

This admittedly doesn't fit with the other trust sub-commands too well,
since the others are all expecting 'mos' tree paths.  That was
problematic for me as I just wanted to generate a cert.pem and
privkey.pem.

But it felt like maybe trust is too inter-twined with mos if it is only
able to read paths that are written by mos.  Should those functions
that expect/enforce directory structure be part of the 'mos' binary and
calling exposed api calls from trust?

Signed-off-by: Scott Moser <smoser@brickies.net>