add support for mosctl install
install-stacker.yaml builds an rfs for the install ISO from which
the install is run.

mos-install is a wrapper script to run the install.

mos-install.service is the systemd script to automatically run it.

live/README.md shows the steps to use this.

All of this, again, should be automated in a 'machine init'.

Signed-off-by: Serge Hallyn <serge@hallyn.com>