Add subcommand "computePCR7" to compute pcr7 for a keyset
This subcommand takes the keysetName as an argument. It
computes the expected pcr7 value based on expected order of tpm
events that occur during a secureboot boot.
Expected uefi secureboot variables, including key databases, and
certificates known to extend into pcr7 during boot are generated
according to tpm specs using the specified keyset and measured to
produce expected pcr7 value.

Should additional drivers and uefi programs be added to the boot, then
this implementation would need to be revisited to include those certificates
that would be extended into pcr7.

The DBX hash and guid is hardcoded from the DBX value in current ovmf_vars.fd
from bootkit. Should this change or when we decide how to manage DBX, this
code will need to be revisited.

Because the mos shim includes 3 uki keys, "computePCR7" returns 3 pcr7 values,
one for each key in the order, pcr7 value, for production, limited, tpm.

Signed-off-by: Joy Latten <jlatten@cisco.com>