Generate tpm policies using go-tpm2 package.
This commit replaces tpm-tools commands with golang
go-tpm2 module to generate tpm ea policies.
Defaults are set for the outputs (policy digest files).
Inputs must be given (the pcr7 files). The policyversion
is anticipated to stay constant for a while so a default
is set for it.

Changelog: 07/26: Serge pushed for a whitespace fix

Signed-off-by: Joy Latten <jlatten@cisco.com>
Signed-off-by: Serge Hallyn <serge@hallyn.com>