Add a 'gofmt' target for make and call it as a trust dep and from c-i
My editor automatically runs 'gofmt'.  So its annoying to open a file,
make a change and see a bunch of whitespace unrelated changes in place.
Ok, maybe I should fix my editor, but really 'gofmt' code is the norm.

The change here is to run gofmt and fail the build if it makes changes.
That way those changes don't get checked in and don't pass c-i if it
hasn't run.

Signed-off-by: Scott Moser <smoser@brickies.net>