bump bootkit layer version
artifacts.go is fetching a bootkit layer to update the artifacts.

Move the definition of it to Makefile, and allow the caller to
specify it.

This is still not how it should be - trust should use the bootkit
api.  But then it should still allow the user to specify the
version.

Signed-off-by: Serge Hallyn <serge@hallyn.com>